feat: jwt를 활용한 CUD 작업 (#66)
* feat: jwt를 활용한 CUD 작업

- 토큰 시간 100분으로 수정
- dto에 time 추가
- 엔티티의 link -> url 변경

* fix: dto, entity 수정

- dto에서 time 제거
- entity에서 time -> duration으로 변경
- != --> !== 로 수정
- jwt 시간 10분으로 수정

+++ src/post/post.service.ts
@@ -1,19 +1,46 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PostCreateDto, PostUpdateDto } from './post.dto';
 import { Post } from './post.entity';
+import { Request } from 'express';
+import { JwtUserDto } from 'src/auth/auth.dto';
+import { UserService } from 'src/user/user.service';
+import { isNil } from 'src/common/utils';
 
 @Injectable()
 export class PostService {
   constructor(
     @InjectRepository(Post)
     private readonly postRepository: Repository<Post>,
+    private readonly userService: UserService,
   ) {}
 
-  create(_postCreateDto: PostCreateDto) {
-    // TODO
-    throw new Error('Method not implemented.');
+  async create(req: Request, postCreateDto: PostCreateDto) {
+    const jwtUserDto = req.user as JwtUserDto;
+    const { provider, providerId } = jwtUserDto;
+
+    const author = await this.userService.findByProviderAndProviderId(
+      provider,
+      providerId,
+    );
+
+    if (isNil(author)) {
+      throw new Error('User Not Exist');
+    }
+
+    const post = new Post();
+    post.title = postCreateDto.title;
+    post.endDate = postCreateDto.endDate;
+    post.gender = postCreateDto.gender;
+    post.ages = postCreateDto.ages;
+    post.researchType = postCreateDto.researchType;
+    post.url = postCreateDto.url;
+    post.procedure = postCreateDto.procedure;
+    post.duration = postCreateDto.duration;
+    post.content = postCreateDto.content;
+    post.author = author;
+    return this.postRepository.save(post);
   }
 
   findAll() {
@@ -23,15 +50,62 @@ export class PostService {
   findOne(id: number) {
     return this.postRepository.findOne({
       where: { id },
+      relations: ['author'],
     });
   }
 
-  update(_id: number, _postUpdateDto: PostUpdateDto) {
-    // TODO
-    throw new Error('Method not implemented.');
+  async update(req: Request, id: number, postUpdateDto: PostUpdateDto) {
+    const jwtUserDto = req.user as JwtUserDto;
+    const { provider, providerId } = jwtUserDto;
+
+    const author = await this.userService.findByProviderAndProviderId(
+      provider,
+      providerId,
+    );
+
+    if (isNil(author)) {
+      throw new Error('User Not Exist');
+    }
+
+    const post = await this.findOne(id);
+
+    if (isNil(post)) {
+      throw new Error('post not found');
+    }
+
+    if (post.author.id !== author.id) {
+      throw new UnauthorizedException('Not the owner of Post');
+    }
+
+    return this.postRepository.update(id, postUpdateDto);
   }
 
-  remove(id: number) {
+  async remove(req: Request, id: number) {
+    const jwtUserDto = req.user as JwtUserDto;
+    const { provider, providerId } = jwtUserDto;
+
+    const author = await this.userService.findByProviderAndProviderId(
+      provider,
+      providerId,
+    );
+
+    if (isNil(author)) {
+      throw new Error('User Not Exist');
+    }
+
+    const post = await this.findOne(id);
+
+    if (isNil(post)) {
+      throw new Error('post not found');
+    }
+
+    console.log(post.author);
+    console.log(author);
+
+    if (post.author.id !== author.id) {
+      throw new UnauthorizedException('Not the owner of Post');
+    }
+
     return this.postRepository.delete(id);
   }
 }
